% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIDFromFilename.R
\name{add.id}
\alias{add.id}
\title{add.id}
\usage{
add.id(indir, outdir, save = TRUE, index.id = c(0, 6))
}
\arguments{
\item{indir}{A string containing the directory where the raw data is from}

\item{outdir}{A string containing the directory where the processed data should be stored (optional, necessary if save = TRUE)}

\item{save}{A logical value which defines whether the generated datasets should be stored or not}

\item{index.id}{A vector containing two integer. The first one marks the start index of the logger id in the filename, the second one the end}
}
\value{
Returns the processed dataset containing logger data
}
\description{
\code{add.id} adds the id inside the filename to the dataframe
}
\examples{

add.id("/home/ela/Documents/R-FinalExam/packagetest/reorder", "/home/ela/Documents/R-FinalExam/packagetest/append_df/", save = TRUE, index.id = c(0,6))
add.id("/home/ela/Documents/R-FinalExam/packagetest/reorder", "/home/ela/Documents/R-FinalExam/packagetest/append_df/", save = FALSE, index.id = c(0,6))
}
