% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coordinates.R
\name{add_coordinates}
\alias{add_coordinates}
\title{add_coordinates}
\usage{
add_coordinates(
  in_file,
  logger_coordinates_file,
  add_coordinates_By_Column,
  csv_sep = ",",
  save = FALSE,
  out_dir = NULL
)
}
\arguments{
\item{in_file}{Type: String; The input file containing the logger data, can be a csv or xlsx file}

\item{logger_coordinates_file}{Type: String; This file has the coordinates for each logger, can be a csv or xlsx file}

\item{add_coordinates_By_Column}{Type: String; The column name of input file and the coordinate dataset on which the two datasets should be merged (Needs to be the same!)}

\item{csv_sep}{Optional Argument, Default Value: ",", Type: String; If one of the datasets is a csv file the seperator needs to be defined, at this moment the seperator needs to be the same in both datasets}

\item{save}{Optional Argument, Default Value: FALSE, Type: Logical; Can be set to TRUE, if the added coordinates sohuld be saved inside the input file}

\item{out_dir}{Optional Argument, Default Value: NULL, Type: String; The variable defines the output directory of the dataset, that should be saved}
}
\value{
Returns a dataframe (logger_df) that has new columns for x,y (,z) added
}
\description{
\code{add_coordinates} does smth.
}
\examples{
add_coordinates(in_file = "~/Documents/R-FinalExam/packagetest/append_df/combined_loggerByID.csv", logger_coordinates_file = "~/Documents/R-FinalExam/Loggerpositionen_Muragl.csv" ,add_coordinates_By_Column ="Logger.ID")
add_coordinates(in_file = "~/Documents/R-FinalExam/packagetest/append_df/combined_loggerByID.csv", logger_coordinates_file = "~/Documents/R-FinalExam/Loggerpositionen_Muragl.csv" , add_coordinates_By_Column = "Logger.ID", csv_sep = "," , save = TRUE, out_dir = "~/Documents/R-FinalExam/packagetest/spatialdata/")

}
