% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIDFromFilename.R
\name{add_id}
\alias{add_id}
\title{add_id}
\usage{
add_id(in_dir, csv_sep = ",", out_dir = NULL, save = FALSE, index_id = c(0, 6))
}
\arguments{
\item{in_dir}{A string containing the directory where the raw data is from}

\item{csv_sep}{A string containg the seperator of the csv file, default value is ",", but can be changed to e.g. ";" , "."#'}

\item{out_dir}{A string containing the directory where the processed data should be stored (optional, necessary if save = TRUE)}

\item{save}{A logical value which defines whether the generated datasets should be stored or not}

\item{index_id}{A vector containing two integer. The first one marks the start index of the logger id in the filename, the second one the end}
}
\value{
Returns the processed dataset containing logger data
}
\description{
\code{add_id} adds the id inside the filename to the dataframe
}
\examples{

add_id("~/Documents/R-FinalExam/packagetest/reorder/", csv_sep = "," , "~/Documents/R-FinalExam/packagetest/append_df/", save = TRUE)
add_id("~/Documents/R-FinalExam/packagetest/reorder/", csv_sep = ",")

}
