% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aggregation.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Aggregate Data}
\usage{
aggregate_data(df, aggregation_type, temperature_column)
}
\arguments{
\item{df}{MANDATORY Type: dataframe- - the input Dataframe containing the Loggerdata (please apply rename colnames and modify dates before)}

\item{aggregation_type}{MANDATORY - Type: String - you can choose between "daily", "monthly", "seasonal" and "annual"}

\item{temperature_column}{OPTIONAL - Type: String - if temperature column is not "Temperature_C" change the parameter}
}
\value{
returns a dataframe with the aggregated temperature data. it will contain new columns for mean, standard derivation, minimum and maximum Temperature
}
\description{
Data aggregation by Day, Month, Season or Year
}
