% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_plot.R
\name{create_daily_plot}
\alias{create_daily_plot}
\title{create_daily_plot}
\usage{
create_daily_plot(data, title)
}
\arguments{
\item{data}{A dataframe containing daily temperature data with columns "Julian", "mean_temperature", and "Logger_ID".}

\item{title}{The title of the plot}
}
\value{
A ggplot object displaying the temperature development over time.
}
\description{
Creates a plot showing the development of temperature over time for a single logger.
}
