% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_plots.R
\name{create_daily_plots}
\alias{create_daily_plots}
\title{create_daily_plots}
\usage{
create_daily_plots(data)
}
\arguments{
\item{data}{uses a list of daily temperature data to create plots by site}
}
\value{
a list of plots
}
\description{
create_daily_plots
}
