% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_plots_by_month.R
\name{create_monthly_plots_by_month}
\alias{create_monthly_plots_by_month}
\title{create_monthly_plots_by_month}
\usage{
create_monthly_plots_by_month(data)
}
\arguments{
\item{data}{a list of dataframes containing daily aggregated datasets}
}
\value{
A list of plots
}
\description{
Creates a List of Plots for each month of a year where the monthly temperature developements for all plots are shown
}
