% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_plots_by_site.R
\name{create_monthly_plots_by_site}
\alias{create_monthly_plots_by_site}
\title{create_monthly_plots_by_site}
\usage{
create_monthly_plots_by_site(data)
}
\arguments{
\item{data}{list of monthly aggregated datasets}
}
\value{
a list of plots
}
\description{
Creates a list of plots, each plot contains an overview over the temperature for the whole time span aggregated by month
}
