% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subset.R
\name{create_subset}
\alias{create_subset}
\title{Create Subset}
\usage{
create_subset(data, columns)
}
\arguments{
\item{data}{The input dataframe.}

\item{columns}{A vector containing the names of columns to be retained in the subset.}
}
\value{
A subset of the original dataframe.
}
\description{
\code{create_subset()} generates a subset of a given dataframe based on a vector of column names.
}
