% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_task_bundle.R
\name{download_task_bundle}
\alias{download_task_bundle}
\title{download data from AppEEARS}
\usage{
download_task_bundle(task_id, token, output_directory)
}
\arguments{
\item{task_id}{MANDATORY - the ID generated by the function submit_processing_task or an id for a task in AppEEARS}

\item{token}{MADATORY - the token generated during the log-in process (appeears_login)}

\item{output_directory}{MANDATORY - the output directory}
}
\value{
returns the status of the download process
}
\description{
download data from AppEEARS
}
