% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{extract_date_time}
\alias{extract_date_time}
\title{extract_date_time}
\usage{
extract_date_time(indir, outdir, save)
}
\arguments{
\item{indir}{A string containing the directory where the raw data is from}

\item{outdir}{A string containing the directory where the processed data should be stored (optional, necessary if save = TRUE)}

\item{save}{A logical value which defines whether the generated datasets should be stored or not}
}
\value{
Returns the processed dataset containing logger data
}
\description{
\code{extract_date_time} does something.
}
\examples{
extract_date_time("/home/ela/Documents/R-FinalExam/Muragl", "/home/ela/Documents/R-FinalExam/reordered_data/", save = TRUE)
extract_date_time("/home/ela/Documents/R-FinalExam/Muragl", "/home/ela/Documents/R-FinalExam/reordered_data/", save = FALSE)


}
