% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoggerToSpatial.R
\name{make_loggerdata_spatial}
\alias{make_loggerdata_spatial}
\title{Title}
\usage{
make_loggerdata_spatial(
  in_file,
  csv_sep = ",",
  input_EPSG,
  output_crs = "+proj=longlat +datum=WGS84",
  save = FALSE,
  out_dir = NULL
)
}
\arguments{
\item{in_file}{Type: String; The input file containing the logger data including columns for coordinates named "X", "Y", can be a csv or xlsx file}

\item{csv_sep}{Optional Argument, Default Value: ",", Type: String, the separator for the in_file vairable}

\item{input_EPSG}{a String - the epsg code for the inputdata - structure: "EPSG:21781"}

\item{output_crs}{optional, a string containing the output crs. Default value is: "+proj=longlat +datum=WGS84"}

\item{save}{Optional Argument, Default Value: FALSE, Type: Logical; Can be set to TRUE, if the added coordinates sohuld be saved inside the input file}

\item{out_dir}{Optional Argument, Default Value: NULL, Type: String; The variable defines the output directory of the dataset, that should be saved}
}
\value{
Returns two gpkg-files, one in the original epsg and one in WGS1984
}
\description{
Title
}
\examples{
make_loggerdata_spatial(in_file = "~/Documents/R-FinalExam/packagetest/spatialdata/logger_geodata.csv", csv_sep = "," , input_EPSG = "EPSG:21781", output_crs = "+proj=longlat +datum=WGS84", save = FALSE, out_dir = NULL)
make_loggerdata_spatial(in_file  = "~/Documents/R-FinalExam/packagetest/spatialdata/logger_geodata.csv", csv_sep = "," , input_EPSG = "EPSG:21781", output_crs = "+proj=longlat +datum=WGS84", save = TRUE, out_dir = "~/Documents/R-FinalExam/packagetest/spatialdata/")

}
