% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spatial.R
\name{make_spatial_data}
\alias{make_spatial_data}
\title{Convert Spatial Data to WGS84 Coordinates}
\usage{
make_spatial_data(
  df,
  coordinate_column = c("X", "Y"),
  data_crs_original,
  transformed_crs = "+proj=longlat +datum=WGS84",
  logger_id_column = "Logger_ID"
)
}
\arguments{
\item{df}{A data frame containing spatial data.}

\item{coordinate_column}{A character vector specifying the column names for the X and Y coordinates.}

\item{data_crs_original}{The CRS of the original spatial data.}

\item{transformed_crs}{The target CRS for transformation. Defaults to WGS84.}

\item{logger_id_column}{The name of the column containing the logger IDs. Defaults to "Logger_ID".}
}
\value{
A list with two elements:
\item{spatial_object}{The spatial object transformed to WGS84 coordinates.}
\item{coordinates_df}{A data frame containing the WGS84 longitude, latitude coordinates along with the logger IDs.}
}
\description{
This function takes spatial data in a specified coordinate reference system (CRS) and transforms it to WGS84 coordinates.
}
