% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_loggerDs.R
\name{merge_logger.ds}
\alias{merge_logger.ds}
\title{merge_logger.ds}
\usage{
merge_logger.ds(indir, outdir, save = TRUE)
}
\arguments{
\item{indir}{A string containing the directory where the raw data is from}

\item{outdir}{A string containing the directory where the processed data should be stored (optional, necessary if save = TRUE)}

\item{save}{A logical value which defines whether the generated datasets should be stored or not}
}
\value{
Returns one csv dataset containing all loggerdata
}
\description{
\code{merge_logger.ds} does something.
}
\examples{

merge_logger.ds(indir = '/home/ela/Documents/R-FinalExam/packagetest/append_df', outdir = '/home/ela/Documents/R-FinalExam/packagetest/append_df/', save = TRUE)
merge_logger.ds(indir = '/home/ela/Documents/R-FinalExam/packagetest/append_df', outdir = '/home/ela/Documents/R-FinalExam/packagetest/append_df/', save = FALSE)

}
