% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_loggerDs.R
\name{merge_logger_ds}
\alias{merge_logger_ds}
\title{merge_logger_ds}
\usage{
merge_logger_ds(in_dir, csv_sep = ",", out_dir = NULL, save = FALSE)
}
\arguments{
\item{in_dir}{A string containing the directory where the raw data is from}

\item{csv_sep}{A string containg the seperator of the csv file, default value is ",", but can be changed to e.g. ";" , "."}

\item{out_dir}{A string containing the directory where the processed data should be stored (optional, necessary if save = TRUE)}

\item{save}{A logical value which defines whether the generated datasets should be stored or not}
}
\value{
Returns one csv dataset containing all loggerdata
}
\description{
\code{merge_logger_ds} does something.
}
\examples{

merge_logger_ds(in_dir = '~/Documents/R-FinalExam/packagetest/append_df/', out_dir = '~/Documents/R-FinalExam/packagetest/append_df/', save = TRUE)
merge_logger_ds(in_dir = '~/Documents/R-FinalExam/packagetest/append_df/')

}
