% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_dates.R
\name{mutate_dates}
\alias{mutate_dates}
\title{Mutate Time Column}
\usage{
mutate_dates(df, datetime_column, time_format)
}
\arguments{
\item{df}{The input dataframe containing time-related columns.}

\item{datetime_column}{The name of the column containing both Date and Time information, or just Date.}

\item{time_format}{The format of the datetime information in the datetime_column. This should follow standard datetime formatting conventions, such as "\%Y-\%m-\%d \%H:\%M:\%S".}
}
\value{
A modified dataframe with additional columns for Date, Julian day, Month, and Year extracted from the datetime_column.
}
\description{
This function mutates the time-related columns in the dataframe to extract useful components such as Date, Julian day, Month, and Year.
}
