% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{read_data
reads in files used for the analysis of logger data. Possible file types are: Excel, csv or a dataframe}
\usage{
read_data(
  input_directory,
  csv_sep = ",",
  csv_comment_character = "#",
  add_ID_from_filename = TRUE,
  index_id = c(0, 6)
)
}
\arguments{
\item{input_directory}{MANDATORY, Type: String - The input directory where the file(s) is/are stored}

\item{csv_sep}{MANDATORY, Type: String - The csv seperator used in the file. If no}

\item{csv_comment_character}{OPTIONAL, Type: String - default values: ","}

\item{add_ID_from_filename}{OPTIONAL, Type: TRUE/FALSE - default value: TRUE}

\item{index_id}{OPTIONAL, Type: vector - should contain the start and end position of the ID in the filename, default: c(0,6)}
}
\value{
a single dataframe or a list of dataframes containing the raw data
}
\description{
read_data
reads in files used for the analysis of logger data. Possible file types are: Excel, csv or a dataframe
}
