% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_colnames.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename Columns}
\usage{
rename_columns(
  df,
  rename_map = list(Number = "No", Logger_ID = "Logger_ID", Time = "Time", Temperature_C
    = "X1.oC", Battery_Voltage = "HK.Bat.V")
)
}
\arguments{
\item{df}{The input dataframe to which the changes should be applied.}

\item{rename_map}{An optional list specifying the old and new column names.
The structure should be: rename_map = list(New_Colname1 = "Old_Colname1", New_Colname2 = "Old_Colname2", ...)
Default: list(Number = "No", Logger_ID = "Logger_ID", Time = "Time", Temperature_C = "X1.oC", Battery_Voltage = "HK.Bat.V").}
}
\value{
A dataframe with the new column names.
}
\description{
Renames columns of a dataframe according to the provided mapping.
}
