% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_appeears_products.R
\name{show_appeears_products}
\alias{show_appeears_products}
\title{Get an overview over all products of AppEEARS}
\usage{
show_appeears_products()
}
\value{
returns a dataframe with all available products, their description and temporal resolution
}
\description{
Shows the list of all appeears products
}
