% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_processing_task.R
\name{submit_processing_task}
\alias{submit_processing_task}
\title{submit a processing task to AppEEARS}
\usage{
submit_processing_task(
  task_name,
  products_df,
  filter_topic,
  token,
  start_date,
  end_date,
  coordinates
)
}
\arguments{
\item{task_name}{MANDATORY -  Type: String - Name of the task you want to download}

\item{products_df}{MANDATORY -  Type: dataframe - created through the function show_appeears_products()}

\item{filter_topic}{MANDATORY -  Type: String - the topic of your dataanalysis (e.g. LST, NDVI, NDSI)}

\item{token}{MANDATORY -  Type:  - String in JSON Type - generated through the function appeears_login()}

\item{start_date}{MANDATORY -  Type: String - in format dd-mm-yy}

\item{end_date}{MANDATORY -  Type: String - in format dd-mm-yy}

\item{coordinates}{MANDATORY -  Type: array/matrix - in longitude, latitude}
}
\value{
task status, task_id and the token
}
\description{
the function "submit_processing_task()" submits a data request to the AppEEARS platform.
}
