% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coordinate_column.R
\name{add_coordinates}
\alias{add_coordinates}
\title{Add Coordinates to the dataframe}
\usage{
add_coordinates(df, logger_coordinates, Coordinate_Column_Name = "Logger_ID")
}
\arguments{
\item{df}{\itemize{
\item Type: Dataframe - The input dataframe to which coordinates will be added.
}}

\item{logger_coordinates}{\itemize{
\item Type: String or Dataframe - The file path (Excel, CSV) or dataframe containing the coordinates.
}}

\item{Coordinate_Column_Name}{\itemize{
\item Type: String - The name of the column in 'logger_coordinates' that contains the identifier (logger ID) for merging.
}}
}
\value{
Dataframe - The input dataframe with added coordinates based on the identifier column.
}
\description{
'add_coordinates' will add the coordinates in new columns to an existing dataframe. They are added based on an identifier column
}
