% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coordinate_column.R
\name{add_coordinates}
\alias{add_coordinates}
\title{Add Coordinates to the dataframe}
\usage{
add_coordinates(df, logger_coordinates, Coordinate_Column_Name = "Logger_ID")
}
\arguments{
\item{df}{MANDATORY - Type: Dataframe - input dataframe}

\item{logger_coordinates}{MANDATORY - Type: String , contains the directory of the file (excel, csv) or dataframe with the coordinates}

\item{Coordinate_Column_Name}{MANDATORY - Type: String - the column that contains the logger id and on which the coordinates will be merged to the dataframe}
}
\value{
logger dataset with the coordinates for the logger position
}
\description{
'add_coordinates' will add the coordinates in new columns to an existing dataframe. They are added based on an identifier column
}
