% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aggregation.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Aggregate Data}
\usage{
aggregate_data(df, aggregation_type, aggregation_column)
}
\arguments{
\item{df}{A dataframe containing the Logger data. Columns should include "Logger_ID", "Month", "Date", "Year", "Julian", and the specified aggregation column.}

\item{aggregation_type}{The type of aggregation to perform: "daily", "monthly", "seasonal", or "annual".}

\item{aggregation_column}{The name of the column to use for aggregation.}
}
\value{
A dataframe with the aggregated data including mean, standard deviation, minimum, and maximum of the specified aggregation column.
}
\description{
Perform data aggregation by Day, Month, Season, or Year based on a specified aggregation column.
}
