% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aggregation.R
\name{aggregate_data}
\alias{aggregate_data}
\title{aggregate_data}
\usage{
aggregate_data(
  input_file,
  column_names = list(id = "ID", date = "Date", year = "Year", month = "Month", time =
    "Time", temperature = "temp"),
  daily = TRUE,
  monthly = TRUE,
  annual = TRUE,
  seasonal = TRUE
)
}
\arguments{
\item{input_file}{Type: String: The input file containing the data to be aggregatet}

\item{column_names}{Optional, Type: List; a list with the column names of the input file. Default is:  list(id = 'ID', date = "Date" , year = "Year", month = "Month", time = "Time", temperature = "temp")}

\item{daily}{Optional, Type: logical; Default value is TRUE}

\item{monthly}{Optional, Type: logical; Default value is TRUE}

\item{annual}{Optional, Type: logical; Default value is TRUE}

\item{seasonal}{Optional, Type: logical; Default value is TRUE}
}
\value{
returns dataframes for aggregatet temperature data (daily, monthly, seasonal, annual)
}
\description{
\code{aggregate_data} aggregates data daily, monthly, seasonal and annual and calculates Mean, Standard Derivation, Minimum and Maximum for the Temperature
}
\examples{
aggregate_data(input_file= "~/Documents/R-FinalExam/packagetest/spatialdata/logger_geodata.csv", column_names = list(id = 'Logger.ID', date = "Date" , year = "Year", month = "Month", time = "Time", temperature = "X1.oC"))
aggregate_data(input_file= "~/Documents/R-FinalExam/packagetest/spatialdata/logger_geodata.csv", column_names = list(id = 'Logger.ID', date = "Date" , year = "Year", month = "Month", time = "Time", temperature = "X1.oC"), daily = TRUE, monthly = TRUE, annual = TRUE, seasonal = TRUE)

}
