% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppEEARS_login.R
\name{appeears_login}
\alias{appeears_login}
\title{AppEEARS login}
\usage{
appeears_login()
}
\value{
returns a token that is needed for further data access
}
\description{
'appeears_login' checks if a netrc file exists,
if not it sets it up and then checks the credentials and logs in to Appears
}
