% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subset.R
\name{create_subset}
\alias{create_subset}
\title{Create Subset}
\usage{
create_subset(data, columns)
}
\arguments{
\item{data}{The input dataframe}

\item{columns}{A vector containing all columns that should stay in the dataframe}
}
\value{
returns a subset of the original dataframe
}
\description{
create_subset() creates a subset of a given dataframe based on a vector of columns
}
