% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_task_bundle.R
\name{download_task_bundle}
\alias{download_task_bundle}
\title{Download data from AppEEARS}
\usage{
download_task_bundle(task_id, token, output_directory)
}
\arguments{
\item{task_id}{The ID of the task generated by the function \code{submit_processing_task} or an ID for a task in AppEEARS.}

\item{token}{The access token generated during the login process (\code{appeears_login}).}

\item{output_directory}{The directory where the downloaded files will be saved.}
}
\value{
Returns the status of the download process.
}
\description{
Downloads data files associated with a specific task from AppEEARS.
}
