% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spatial.R
\name{make_spatial_data}
\alias{make_spatial_data}
\title{Create a spatial object in wgs84 coordinates}
\usage{
make_spatial_data(
  df,
  coordinate_column = c("X", "Y"),
  data_crs_original,
  transformed_crs = "+proj=longlat +datum=WGS84",
  preferred_spatial_object = "logger_wgs84"
)
}
\arguments{
\item{df}{MANDATORY - Type:dataframe - df must be with coordinates}

\item{coordinate_column}{OPTIONAL - Type: a vector - defining the columnnames for the coordinates}

\item{data_crs_original}{MANDATORY - Type: String - String with EPSG code in the format :"EPSG:XXXXX"}

\item{transformed_crs}{OPTIONAL - Type: String - the crs to which the original should be transformed}

\item{preferred_spatial_object}{OPTIONAL - Type: String - choose between : "logger_wgs84" , "logger_original_crs"}
}
\value{
an sf object
}
\description{
Create a spatial object in wgs84 coordinates
}
