% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_dates.R
\name{mutate_dates}
\alias{mutate_dates}
\title{mutate time column}
\usage{
mutate_dates(df, time_column, time_format)
}
\arguments{
\item{df}{MANDATORY, Dataframe containing the time-related columns}

\item{time_column}{MANDATORY, Type: String, The column name containing the Date and Time or only Date column}

\item{time_format}{MANDATORY, Type: String, The format of Time and Date. For example: "\%d.\%m.\%Y \%H:\%M:\%S"}
}
\value{
dataframe with the correct Time, Date, Juliam Month and Year columns
}
\description{
This function mutates the time-related columns in the dataframe, including Time, Date,
Julian, Month, and Year.
}
\examples{

mutated_df <- mutate_dates(df = read.csv("/home/ela/Documents/R-FinalExam/examples/02_logger.csv"), time_column = "Time", time_format = "\%d.\%m.\%Y \%H:\%M:\%S")

}
