% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_colnames.R
\name{rename_columns}
\alias{rename_columns}
\title{rename_columns}
\usage{
rename_columns(
  df,
  rename_map = list(Number = "No", Logger_ID = "Logger_ID", Time = "Time", Temperature_C
    = "X1.oC", Battery_Voltage = "HK.Bat.V")
)
}
\arguments{
\item{df}{MANDADORY - the input dataframe you want to apply the changes to}

\item{rename_map}{OPTIONAL - a list containing all old and new column names. The structure is: rename_map = list(New_Colname1 = "Old_Colname1, New_Colname2 = "Old_Colname2", ...)
default: list(Number = "No", Logger_ID = "Logger_ID", Time = "Time", Temperature_C = "X1.oC", Battery_Voltage = "HK.Bat.V")}
}
\value{
returns a dataframe with the new column names
}
\description{
rename_columns
}
\examples{
list_df <- read_data(input_directory = "/home/ela/Documents/R-FinalExam/Muragl/")
logger1 <- list_df$A50276_20231006143640.csv


# Example renaming map

}
